Replate trait with new service. Use DI.
